# 🚀 Release v1.1.0: Major Performance Optimizations & Security Fixes