--- a/ansible/ansible.cfg
+++ b/ansible/ansible.cfg
@@ -3,6 +3,7 @@ inventory = ./inventory/tofu_inventory.py
 host_key_checking = False
 deprecation_warnings = False
 remote_user = abevz
+stdout_callback = yaml
 
 [privilege_escalation]
 become = True