--- a/ansible/templates/containerd-config.toml.j2
+++ b/ansible/templates/containerd-config.toml.j2
@@ -2,7 +2,7 @@ version = 2
 
 [plugins]
   [plugins."io.containerd.grpc.v1.cri"]
-    sandbox_image = "registry.k8s.io/pause:3.10" # Используем версию 3.10, как рекомендовано в логах
+    sandbox_image = "registry.k8s.io/pause:3.10" # Using version 3.10 as recommended in the logs
     [plugins."io.containerd.grpc.v1.cri".containerd]
       default_runtime_name = "runc"
       [plugins."io.containerd.grpc.v1.cri".containerd.runtimes]
@@ -11,6 +11,6 @@ version = 2
           [plugins."io.containerd.grpc.v1.cri".containerd.runtimes.runc.options]
             SystemdCgroup = true
 
-    # Новая, правильная конфигурация для registry
+    # New, correct registry configuration
     [plugins."io.containerd.grpc.v1.cri".registry]
       config_path = "/etc/containerd/certs.d"