--- a/ansible/templates/hosts_harbor.toml.j2
+++ b/ansible/templates/hosts_harbor.toml.j2
@@ -1,0 +1,9 @@
+# See https://github.com/containerd/containerd/blob/main/docs/hosts.md
+server = "{{ upstream_server }}"
+
+[host."https://{{ harbor_hostname }}"]
+  capabilities = ["pull", "resolve"]
+  override_path = true
+  
+[host."https://{{ harbor_hostname }}".header]
+Authorization = ["Basic {{ harbor_auth }}"]