--- a/config.conf
+++ b/config.conf
@@ -6,7 +6,7 @@
 
 # --- Core Configuration ---
 CPC_ENV_FILE="cpc.env"
-CLUSTER_CONTEXT_FILE=".cluster_context"
+CPC_CONTEXT_FILE=".cluster_context"
 REPO_PATH=""  # Will be set dynamically by setup-cpc
 
 # --- Color definitions for output ---
@@ -58,5 +58,5 @@ DOCS_DIR="docs"
 SCRIPTS_DIR="scripts"
 
 # --- Export commonly used variables ---
-export CPC_ENV_FILE CLUSTER_CONTEXT_FILE
+export CPC_ENV_FILE CPC_CONTEXT_FILE
 export RED GREEN YELLOW BLUE PURPLE CYAN WHITE ENDCOLOR