--- a/config.conf
+++ b/config.conf
@@ -57,6 +57,33 @@ ENVIRONMENTS_DIR="envs"
 DOCS_DIR="docs"
 SCRIPTS_DIR="scripts"
 
+# --- Error Handling Configuration ---
+# Timeout values in seconds
+command_timeout=300      # 5 minutes for general commands
+network_timeout=30       # 30 seconds for network operations
+ansible_timeout=1800     # 30 minutes for Ansible operations
+kubectl_timeout=120      # 2 minutes for kubectl operations
+terraform_timeout=3600   # 1 hour for Terraform/OpenTofu operations
+
+# Retry configuration
+max_retries=3
+base_retry_delay=2
+max_retry_delay=60
+retry_backoff_multiplier=2
+
+# Error handling severity levels
+# CRITICAL: Stop execution immediately
+# HIGH: Log error and continue with warnings
+# MEDIUM: Log error and attempt recovery
+# LOW: Log warning only
+# INFO: Log info only
+default_error_severity=MEDIUM
+
+# Recovery configuration
+enable_recovery=true
+recovery_log_path="/tmp/cpc_recovery.log"
+auto_rollback_on_failure=true
+
 # --- Export commonly used variables ---
 export CPC_ENV_FILE CPC_CONTEXT_FILE
 export RED GREEN YELLOW BLUE PURPLE CYAN WHITE ENDCOLOR