… .github/ISSUE_TEMPLATE/bug_report.yml
… .github/ISSUE_TEMPLATE/documentation.yml
… .github/ISSUE_TEMPLATE/feature_request.yml
… .github/ISSUE_TEMPLATE/question.yml
… .github/pull_request_template.md
… CHANGELOG.md
… CONTRIBUTING.md
… README.md
… RELEASE_NOTES.md
… RELEASE_PREPARATION.md
… ansible/playbooks/initialize_kubernetes_cluster_with_dns.yml
… ansible/playbooks/install_kubernetes_cluster.yml
… ansible/playbooks/main.yml
… ansible/playbooks/pb_add_nodes.yml
… ansible/playbooks/pb_delete_node.yml
… ansible/playbooks/pb_prepare_node.yml
… ansible/playbooks/pb_upgrade_addons_extended.yml
… ansible/playbooks/traefik-values.yaml
… ansible/playbooks/validate_cluster.yml
… ansible/roles/setup_kubernetes_cluster/defaults/main.yml
… bashtest/README.md
… bashtest/bash_test_framework.sh
… bashtest/test_ansible_module.sh
… bashtest/test_core_module.sh
… bashtest/test_k8s_cluster_module.sh
… bashtest/test_proxmox_module.sh
… bashtest/test_tofu_module.sh
… cpc
… cpc.env
… cpc.env.example