# 🚀 Release v1.1.0: Major Performance Optimizations & Security Fixes

## 📋 Summary
This PR introduces significant performance improvements to the CPC cluster management tool, with cluster-info command optimized from 22+ seconds to under 0.5 seconds, plus critical security fixes for Kubernetes version pinning.

## ✨ New Features
- **cluster-info --quick mode**: Ultra-fast cluster status (0.1s execution time)
- **Two-tier terraform caching**: Short-term (30s) and long-term (5min) cache layers  
- **Smart workspace detection**: Avoids unnecessary terraform workspace switches
- **Context-aware cache management**: Separate cache files per workspace

## 🔒 Security Fixes
- **Pinned Kubernetes versions**: Fixed high-severity issue where kubelet, kubeadm, kubectl versions weren't pinned
- **Version consistency**: Prevents automatic patch updates that could cause cluster instabilities
- **Role defaults**: Changed from 'latest' to specific pinned versions for production safety

## ⚡ Performance Improvements
| Command | Before | After | Improvement |
|---------|--------|-------|-------------|
| `cluster-info` (first run) | 22s | 7.2s | **3x faster** |
| `cluster-info` (cached) | 22s | 0.44s | **50x faster** |
| `cluster-info --quick` | N/A | 0.1s | **220x faster** |

## 🧪 Testing
- ✅ All tests passing (100% success rate)
- ✅ Comprehensive test suite with 59 tests
- ✅ Performance benchmarking validated
- ✅ No breaking changes - fully backward compatible

## 🔧 Technical Changes
- **Optimized terraform operations**: Smart workspace state management
- **Enhanced caching strategy**: Multi-level cache with intelligent invalidation
- **Reduced I/O operations**: Better cache file handling
- **Network efficiency**: Fewer remote state API calls
- **Security hardening**: Kubernetes component version pinning

## 🔧 Code Quality Improvements
- **Magic number elimination**: Replaced hardcoded values with named constants in terraform
- **Hostname collision prevention**: Added mandatory RELEASE_LETTER to all environments
- **Code consistency**: Enhanced error handling and validation in scripts

## 📚 Documentation Updates
- Updated CHANGELOG.md with detailed performance metrics
- Enhanced RELEASE_NOTES.md with v1.1.0 changes
- Updated help text to include --quick option
- Added performance benchmarks

## 🔄 Migration
- No migration needed - all existing commands work as before
- New `--quick` flag available for ultra-fast cluster information
- Kubernetes versions now properly pinned for consistency

## 🎯 Ready for Release
- [x] Version bumped to 1.1.0
- [x] All tests passing
- [x] Documentation updated
- [x] Performance benchmarks validated
- [x] Security fixes applied
- [x] Code review feedback addressed
- [x] Russian comments translated to English
- [x] Magic numbers replaced with constants
- [x] No breaking changes